localStorage Reference error fix

+++ adminsight-client/src/app/services/local-storage.service.ts
@@ -1,23 +1,36 @@
 import { Injectable } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { PLATFORM_ID, Inject } from '@angular/core';
 
 @Injectable({
   providedIn: 'root'
 })
 export class LocalStorageService {
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) { }
+
   get(key: string): any {
-    const value = localStorage.getItem(key);
-    return value ? JSON.parse(value) : null;
+    if (isPlatformBrowser(this.platformId)) {
+      const value = localStorage.getItem(key);
+      return value ? JSON.parse(value) : null;
+    }
+    return null;
   }
 
   set(key: string, value: any): void {
-    localStorage.setItem(key, JSON.stringify(value));
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.setItem(key, JSON.stringify(value));
+    }
   }
 
   remove(key: string): void {
-    localStorage.removeItem(key);
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.removeItem(key);
+    }
   }
 
   clear(): void {
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
   }
 }
